+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added a new explosion sound for Angry Creepers. It's now the default
     * The fuse sound is vanilla but the explosion sound is the old minecraft beta one
+* Configurable Animal flee speed
 
 ## 2.4.10
 * Removed `use_follow_range_changes` entity type tag